backend/requirements.txt
Flask
pyodbc
Flask-Cors
passlib
python-dotenv
waitress
PyJWT